refactor: split keymap domains
